Add support for llvm coverage.
Right now clang versions 14-19 with profile version 8,9,10 are added.
Clean master branch in x86 platform is not linkable with any version of clang. just compilable.
Clean master branch in ARM platform is not compilable with any version of clang. Only in clang-19 if we disable this flag
`-mgeneral-regs-only`it will succesfuly compile and link.
We tested compilability of our patch in above senario and compiled with clang 14-19 in x86 and in ARM it compiled and
linked with clang-19 in the described situation.